rights checks for products

+++ BE/https_server/restapi/mh_products.py
@@ -1,11 +1,15 @@
-from flask import Response
-from flask import request
+from flask import Response, request, abort
 import json, dbhandler as db
 import utility
+from . import authentication as auth
 
 
 @utility.add_required_headers
-def products_GET(**kwargs):    
+def products_GET(**kwargs):
+    user = auth.authenticate()
+    if not user.can_view_products():
+        abort(403)
+
     rows = db.list_products(**kwargs);
     list = [utility.row_to_json(row) for row in rows]
 
@@ -14,6 +18,10 @@ def products_GET(**kwargs):
 
 @utility.add_required_headers
 def products_POST(**kwargs):
+    user = auth.authenticate()
+    if not user.can_create_products():
+        abort(403)
+
     if request.content_type != 'application/json':
         abort(415, utility.ERR_FMTS['BAD_MIME']%'application/json')
 
@@ -23,6 +31,10 @@ def products_POST(**kwargs):
 
 @utility.add_required_headers
 def product_details_GET(**kwargs):
+    user = auth.authenticate()
+    if not user.can_view_products():
+        abort(403)
+
     row = db.get_product(**kwargs);
     object = utility.row_to_json(row)
 
@@ -31,6 +43,10 @@ def product_details_GET(**kwargs):
 
 @utility.add_required_headers
 def product_details_PATCH(**kwargs):
+    user = auth.authenticate()
+    if not user.can_create_products():
+        abort(403)
+
     if request.content_type != 'application/json':
         abort(415, utility.ERR_FMTS['BAD_MIME']%'application/json')
 
@@ -40,12 +56,20 @@ def product_details_PATCH(**kwargs):
 
 @utility.add_required_headers
 def product_details_DELETE(**kwargs):
+    user = auth.authenticate()
+    if not user.can_create_products():
+        abort(403)
+
     db.delete_product(**kwargs)
     return Response()
 
 
 @utility.add_required_headers
 def product_parts_GET(**kwargs):
+    user = auth.authenticate()
+    if not user.can_view_products():
+        abort(403)
+
     rows = db.list_product_parts(**kwargs);
 
     list = [utility.row_to_json(row) for row in rows]
@@ -59,6 +83,10 @@ def product_parts_GET(**kwargs):
 
 @utility.add_required_headers
 def product_parts_POST(**kwargs):
+    user = auth.authenticate()
+    if not user.can_create_products():
+        abort(403)
+
     if request.content_type != 'application/json':
         abort(415, utility.ERR_FMTS['BAD_MIME']%'application/json')
 
@@ -68,6 +96,10 @@ def product_parts_POST(**kwargs):
 
 @utility.add_required_headers
 def product_part_details_GET(**kwargs):
+    user = auth.authenticate()
+    if not user.can_view_products():
+        abort(403)
+
     row = db.get_product_part(**kwargs);
     object = utility.row_to_json(row)
 
@@ -80,6 +112,10 @@ def product_part_details_GET(**kwargs):
 
 @utility.add_required_headers
 def product_part_details_PATCH(**kwargs):
+    user = auth.authenticate()
+    if not user.can_create_products():
+        abort(403)
+
     if request.content_type != 'application/json':
         abort(415, utility.ERR_FMTS['BAD_MIME']%'application/json')
 
@@ -89,12 +125,20 @@ def product_part_details_PATCH(**kwargs):
 
 @utility.add_required_headers
 def product_part_details_DELETE(**kwargs):
+    user = auth.authenticate()
+    if not user.can_create_products():
+        abort(403)
+
     db.delete_product_part(**kwargs)
     return Response()
 
 
 @utility.add_required_headers
 def product_tickets_GET(**kwargs):
+    user = auth.authenticate()
+    if not user.can_view_products() or not user.can_view_tickets():
+        abort(403)
+
     rows = db.list_product_tickets(**kwargs);
     list = [utility.row_to_json(row) for row in rows]
 
@@ -103,6 +147,10 @@ def product_tickets_GET(**kwargs):
 
 @utility.add_required_headers
 def product_part_tickets_GET(**kwargs):
+    user = auth.authenticate()
+    if not user.can_view_products() or not user.can_view_tickets():
+        abort(403)
+
     rows = db.list_product_part_tickets(**kwargs);
     list = [utility.row_to_json(row) for row in rows]
